Intent: centralize edge cache helpers Operators involved: [Coupling|Resonance] Affected invariants: [] Key changes: - move edge cache helpers into tnfr.cache and integrate with EdgeCacheManager - update imports across code, tests, and benchmarks to use tnfr.cache helpers - remove the obsolete tnfr.helpers.edge_cache module Expected risks/dissonances: minimal regression if caches fail to invalidate Metrics: expect unchanged C(t), Si, νf, phase coherence Equivalence map: tnfr.helpers.edge_cache.* -> tnfr.cache.*
