+++ src/tnfr/node.py
@@ -15,7 +15,8 @@ from .alias import (
     set_dnfr,
     set_theta,
 )
-from .helpers import cached_node_list, increment_edge_version, ensure_node_offset_map
+from .cache import increment_edge_version
+from .helpers import cached_node_list, ensure_node_offset_map
 from .graph_utils import supports_add_edge
 from .node_base import NodeBase
 from .locking import get_lock
