+++ src/tnfr/operators/remesh.py
@@ -8,9 +8,9 @@ from io import StringIO
 from collections import deque
 from statistics import fmean, StatisticsError
 
+from ..cache import edge_version_update
 from ..constants import DEFAULTS, REMESH_DEFAULTS, get_aliases, get_param
 from ..helpers.numeric import kahan_sum_nd
-from ..helpers import edge_version_update
 from ..alias import get_attr, set_attr
 from ..rng import make_rng
 from ..callback_utils import CallbackEvent, callback_manager
