+++ tests/test_neighbors_map_cache.py
@@ -1,7 +1,7 @@
 from types import MappingProxyType
 
 from tnfr.metrics.common import ensure_neighbors_map
-from tnfr.helpers.edge_cache import increment_edge_version
+from tnfr.cache import increment_edge_version
 
 
 def test_neighbors_map_reuses_proxy(graph_canon):
